[AB#1669514] move type to test file so it's not bundled
