+++ uuid.go
@@ -4,7 +4,6 @@ package gosnowflake
 
 import (
 	"crypto/rand"
-	"database/sql/driver"
 	"fmt"
 	"strconv"
 )
@@ -48,61 +47,4 @@ func ParseUUID(str string) UUID {
 
 func (u UUID) String() string {
 	return fmt.Sprintf("%x-%x-%x-%x-%x", u[0:4], u[4:6], u[6:8], u[8:10], u[10:])
-}
-
-// This is for unit testing scans/value of UUIDs being inserted/read to/from the DB - not intended for public use
-type testUUID = UUID
-
-func newTestUUID() testUUID {
-	return testUUID(NewUUID())
-}
-
-func parseTestUUID(str string) testUUID {
-	return testUUID(ParseUUID(str))
-}
-
-// Scan implements sql.Scanner so UUIDs can be read from databases transparently.
-// Currently, database types that map to string and []byte are supported. Please
-// consult database-specific driver documentation for matching types.
-func (uuid *testUUID) Scan(src interface{}) error {
-	switch src := src.(type) {
-	case nil:
-		return nil
-
-	case string:
-		// if an empty UUID comes from a table, we return a null UUID
-		if src == "" {
-			return nil
-		}
-
-		// see Parse for required string format
-		u := ParseUUID(src)
-
-		*uuid = testUUID(u)
-
-	case []byte:
-		// if an empty UUID comes from a table, we return a null UUID
-		if len(src) == 0 {
-			return nil
-		}
-
-		// assumes a simple slice of bytes if 16 bytes
-		// otherwise attempts to parse
-		if len(src) != 16 {
-			return uuid.Scan(string(src))
-		}
-		copy((*uuid)[:], src)
-
-	default:
-		return fmt.Errorf("Scan: unable to scan type %T into UUID", src)
-	}
-
-	return nil
-}
-
-// Value implements sql.Valuer so that UUIDs can be written to databases
-// transparently. Currently, UUIDs map to strings. Please consult
-// database-specific driver documentation for matching types.
-func (uuid testUUID) Value() (driver.Value, error) {
-	return uuid.String(), nil
 }
